# perf(thrift): simplify skipdecoder tpl code